chore: remove the dev script that no-one uses
The idea was to improve startuptime by letting you choose which services
you wanted to start. However some always needed to be started (like DB
and Redis) and the script didn't have a list of stuff always to start.

I never ended up really using it, and neither did anyone else on the team.

If we want such a script in the future I think its probably better to
write it in Node.js using a library like execa (by sindresorhus) since
large Bash scripts tends to become really unmaintanable.