chore: cleanup dialogporten types generated
At the start we generated types for the REST api based on OpenAPI spec.
We can remove that now since we exclusivly uses GraphQL.

We also had type generation for GraphQL setup, but those types where
never used by the BFF. The BFF currently just needs the GraphQL schema
to be able to stitch it together. We can simply find this commit and
revert part of it if we in the future need generated types for GraphQL.
This could be the case if we want to not just naivly stitch the schema
but do some data proccessing and transformations first.

I think its good for now to cleanup what we don't really need or use.