chore: editorconfig setup
EditorConfig is a config standard for defining if you should use tabs or
spaces and for which files. They have plugins for all known code
editors.

The Dialogporten "backend" uses it and sets their default to 4 spaces.
By having a config in our repo as well developers switching between
projects get the correct spacing in both projects.